Fix negative key error handling
Patchwork-id: 46039
O-Subject: [RHEL6 PATCH] Fix negative key error handling [BZ 806393]
Bugzilla: 806393
RH-Acked-by: Kyle McMartin <kmcmarti@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Fix negative key error handling in the backport of upstream keys patches to
RHEL-6 [BZ 806393].

The problem is a combination of two bits of code:

 (1) In key_reject_and_link() there is:

	key->type_data.x[0] = -error;

     which negates a 32-bit signed integer and assigns it to a 64-bit unsigned
     integer - which does not cause the signed integer to be sign-extended, but
     rather the upper 32-bits are cleared:

	0x00000000000003d6 <+141>:   neg    %r13d
	0x00000000000003d9 <+144>:   mov    %r13d,%r13d
	0x00000000000003dc <+147>:   mov    %r13,0xc0(%rbx)

 (2) And then in keyring_search_aux() there is:

	err = (long)key->type_data.x[0];

     which converts the 64-bit unsigned type to a 64-bit signed type (which
     doesn't actually do anything to the value) and sticks it in a long-type
     variable - which then gets cast to a pointer.  Thus the value is not
     recognised as an error code.

The fix is to make sure the error gets sign-extended in key_reject_and_link().

The problem results in a paging request fault that looks something like this:

BUG: unable to handle kernel paging request at 00000000fffffff2
IP: [<ffffffff8120d3d1>] wait_for_key_construction+0x31/0x80
PGD 0
Oops: 0000 [#1] SMP
last sysfs file: /sys/module/sunrpc/initstate
CPU 0
Modules linked in: nfs lockd fscache nfs_acl auth_rpcgss ip6table_filter
ip6_tables ebtable_nat ebtables ipt_MASQUERADE iptable_nat nf_nat
nf_conntrack_ipv4 nf_defrag_ipv4 xt_state nf_conntrack ipt_REJECT xt_CHECKSUM
iptable_mangle iptable_filter ip_tables autofs4 sunrpc cpufreq_ondemand
powernow_k8 freq_table mperf bridge stp llc ipv6 vhost_net macvtap macvlan tun
kvm_amd kvm tg3 microcode serio_raw amd64_edac_mod edac_core edac_mce_amd
k10temp e100 mii sg i2c_piix4 shpchp ext4 mbcache jbd2 sr_mod cdrom sd_mod
crc_t10dif ata_generic pata_acpi pata_serverworks sata_svw radeon ttm
drm_kms_helper drm i2c_algo_bit i2c_core dm_mirror dm_region_hash dm_log dm_mod
[last unloaded: scsi_wait_scan]

Pid: 3057, comm: mount.nfs Not tainted 2.6.32-254.el6.x86_64 #1 empty
empty/S3992
RIP: 0010:[<ffffffff8120d3d1>]  [<ffffffff8120d3d1>]
wait_for_key_construction+0x31/0x80
RSP: 0018:ffff88083575f648  EFLAGS: 00010246
RAX: ffffffff8120d440 RBX: 00000000ffffff82 RCX: 0000000000000002
RDX: ffffffff8120d420 RSI: 0000000000000000 RDI: 00000000ffffff82
RBP: ffff88083575f658 R08: 0000000000000000 R09: 0000000000000000
R10: 0000000000000001 R11: 0000000000000000 R12: ffff88042e5e34e0
R13: ffffffffa0545e99 R14: ffff88042e5e34e0 R15: ffff88042e5e34e3
FS:  00007f2189d4b700(0000) GS:ffff880028200000(0000) knlGS:0000000000000000
CS:  0010 DS: 0000 ES: 0000 CR0: 000000008005003b
CR2: 00000000fffffff2 CR3: 000000042e671000 CR4: 00000000000006f0
DR0: 0000000000000000 DR1: 0000000000000000 DR2: 0000000000000000
DR3: 0000000000000000 DR6: 00000000ffff0ff0 DR7: 0000000000000400
Process mount.nfs (pid: 3057, threadinfo ffff88083575e000, task
ffff880832948ae0)
Stack:
 ffff88042e5e34e0 00000000ffffff82 ffff88083575f6a8 ffffffff8120de05
<d> 0000000000000000 ffff8808000000d0 ffff88083575f828 000000000000000f
<d> ffffffffa0545eac 000000000000000f ffff8808339c10c0 0000000000000014
Call Trace:
 [<ffffffff8120de05>] request_key+0x65/0xa0
 [<ffffffffa0538ad5>] nfs_idmap_request_key+0xc5/0x170 [nfs]
 [<ffffffffa0538d94>] nfs_idmap_lookup_id+0x34/0x80 [nfs]
 [<ffffffffa0538ef5>] nfs_map_name_to_uid+0x75/0xa0 [nfs]
 [<ffffffffa0532188>] decode_getfattr_attrs+0xbb8/0xf90 [nfs]
 [<ffffffffa05325e3>] decode_getfattr+0x83/0xe0 [nfs]
 [<ffffffffa0532df0>] ? nfs4_xdr_dec_getattr+0x0/0xa0 [nfs]
 [<ffffffffa0532e7f>] nfs4_xdr_dec_getattr+0x8f/0xa0 [nfs]
 [<ffffffffa0413184>] rpcauth_unwrap_resp+0x84/0xb0 [sunrpc]
 [<ffffffffa0532df0>] ? nfs4_xdr_dec_getattr+0x0/0xa0 [nfs]
 [<ffffffffa04078e3>] call_decode+0x1b3/0x800 [sunrpc]
 [<ffffffff810909a0>] ? wake_bit_function+0x0/0x50
 [<ffffffffa0407730>] ? call_decode+0x0/0x800 [sunrpc]
 [<ffffffffa0411e77>] __rpc_execute+0x77/0x350 [sunrpc]
 [<ffffffff81090827>] ? bit_waitqueue+0x17/0xd0
 [<ffffffffa04121b1>] rpc_execute+0x61/0xa0 [sunrpc]
 [<ffffffffa0408365>] rpc_run_task+0x75/0x90 [sunrpc]
 [<ffffffffa0408482>] rpc_call_sync+0x42/0x70 [sunrpc]
 [<ffffffffa0526012>] _nfs4_call_sync+0x22/0x30 [nfs]
 [<ffffffffa0520c98>] _nfs4_proc_getattr+0xa8/0xc0 [nfs]
 [<ffffffffa052339e>] nfs4_proc_getattr+0x4e/0x70 [nfs]
 [<ffffffffa050a21f>] ? nfs_alloc_fattr+0x2f/0xe0 [nfs]
 [<ffffffffa05078cb>] nfs4_get_root+0x7b/0x200 [nfs]
 [<ffffffffa050d2c1>] nfs4_remote_get_sb+0x101/0x200 [nfs]
 [<ffffffff8117becb>] vfs_kern_mount+0x7b/0x1b0
 [<ffffffffa050e0ff>] nfs_do_root_mount+0x7f/0xb0 [nfs]
 [<ffffffffa050e242>] nfs4_try_mount+0x52/0xd0 [nfs]
 [<ffffffffa050ea0a>] nfs_get_sb+0x43a/0x8f0 [nfs]
 [<ffffffff8117becb>] vfs_kern_mount+0x7b/0x1b0
 [<ffffffff8117c072>] do_kern_mount+0x52/0x130
 [<ffffffff8119a6f2>] do_mount+0x2d2/0x8d0
 [<ffffffff811980c2>] ? copy_mount_options+0xf2/0x1a0
 [<ffffffff8119ad80>] sys_mount+0x90/0xe0
 [<ffffffff8100b0f2>] system_call_fastpath+0x16/0x1b

+++ security/keys/key.c
@@ -543,7 +543,7 @@ int key_reject_and_link(struct key *key,
 		atomic_inc(&key->user->nikeys);
 		set_bit(KEY_FLAG_NEGATIVE, &key->flags);
 		set_bit(KEY_FLAG_INSTANTIATED, &key->flags);
-		key->type_data.x[0] = -error;
+		key->type_data.x[0] = -(long)error;
 		now = current_kernel_time();
 		key->expiry = now.tv_sec + timeout;
 		key_schedule_gc(key->expiry + key_gc_delay);
